Rename several Python files for clarity
